Render different layout for admin and common user
If user is admin, he will have Approve Cars option in his navbar,
which allows him to approve or reject incoming AddNewCar request.
On the other hand, if we are dealing with common user, we should
block this option for him, because otherwise application with log 405 error in the
console.